rename camelcase to snakecase
